requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
spacy==3.7.4
cryptography==43.0.0
pydantic==2.9.0
loguru==0.7.2
